requirements.txt
asyncmy==0.2.9
awslambdaric==2.0.12
confluent-kafka==2.4.0
fastapi==0.111.0
joblib==1.4.2
mangum==0.17.0
motor==3.4.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
scikit-learn==1.2.2
SQLAlchemy==2.0.31
uvicorn==0.30.1
